[fix][nonequilibrium][tests] fix gradient calculation in nonequilibrium source calculaiton.
This fixes a bug introduced with !1517 and thus fixes the
test_mpnc_kinetic test. However the newly introduced 2p2c test
references were created with the buggy code and thus have to updated (in
this patch). We make it more robust by increasing the Newton accuracy,
introducing checkpoints and incresing the initial time step size, which
results in many less time steps.