[disc] add implementation of l2-projector
Functionality to create l2 projectors that project the solution
living in a function space basis on a domain grid into the function
space basis living on a target grid. Currently, the algorithm
assumes the grids to be non-conforming and expects a glue object.
An overload for different bases operating on the same grid can
be provided in the future, however, the current algorithm would
work as well (with the overhead of computing trivial intersections).