[test] Disable regression test for 2p2c adaptive sequential
With our current test tool it is very hard to create regression tests
for adaptive grids. The adaption criterion is usually based on a
threshold, so it might be slightly different on different architectures.
This will currently make the test fail, altough physically the solution
might not be very different. For such cases tests are very complicated.
One possibility might be the projection on a known grid on which the
regression error can be computed.

Until such tools are avaiable, we simply compile and run the test but
do not check for regression in the solution.