[projector] outsource projection matrices assembly
This way, a user can get access to the matrices underlying an
l2-projection. The projectors do not allow access to these matrices and
internally use a reduced matrix in case the target domain occupies a
larger geometric region than the origin domain, where the rows and
columns that correspond to dofs not taking part in the projection are
deleted.