[swe][frictionlaws] Use std::vector<FrictionLaw>
For each element a object of a child class of FrictionLaw is
instantiated and stored within a vector.
This makes it possible to change the friction law on runtime, use also friction
laws which have more than one parameter, use spatial variying friction
values and even to use muliple friction laws.