Revert "[facet][tracer] adapt reference sol to improved coupling flux computation"
This reverts commit eb294caf0f876b77de9d02ed19d7d86d8ce19000.
Therein, the reference solutions for the box-facet-coupling tracer tests
were changed after modifying the coupling flux assembly in the scheme.
This however only affected cases of tensorial permeabilities and
shouldn't have influenced the solution of the tracer test.
With the fixes on this branch the old reference solution can be reused.