[discretization] workaround for FVGridGeometry deprecation and clang
If clang is used, false positive deprecation warnings are emitted in
context of the change from FVGridGeometry to GridGeometry. As a
workaround, place ignore pragmas around the corresponding declarations
in propertysystem.hh. This omits all deprecation warnings that might
be triggered by using `GetPropType`. Add a general message which warns a
clang user about this issue and a switch to turn off that message.