[discretization] Implement radially symmetric grid geometries
This implements wrappers for scvs and scvf that turn them into
radially symmetric counterparts by overloading the volume() and
area() functions. We support three modes:

* disc (annulus): A 1d grid is extruded into an annulus or disc
In the case of a disc (inner pos: 0) the inner scvf.area() is 0!

* ball (shell): A 1d grid is extruded into a spherical shell or a ball
In the case of a ball the inner scvf.area() is 0!

* toroid: A 2d grid is extruded into a toroid
In the case that on grid side is aligned with the axis the inner
scvf.area() for all inner scvfs is 0!
The grid points all have to be >= 0!
We also support unstructured grids and random convex shapes

(cherry picked from commit bdfd233796ff034bc51e1095e9ca52b00752729a)