[material][co2] Add script to extract co2 tables using NIST webbook
The webbook is also based on the Span Wagner EOS we used for
generating the tables so far. As the original Fortran code
provided by Span and Wagner is not published under an open-source
license, we cannot include in the Dumux source code.
This provided a script that uses NIST's free webbook online
service to generate the tables. The tables are updates because
there seem to be very small differences, possibly due to
differences in rounding error in the computation in the background.