Added a generic linear backend using the solver factories from istl.
This requires a version newer than 2.7.0. The backend translates
the configuration parameters from Dumux CamelCase to the dune-istl
names. If a Dumux parameter is not found we fall back to using
the dune-istl parameter name.

We fall back to the parallel AMG backend if the DUNE version is
not sufficient.