[test] Remove Maxwell-Stefan tracer test
The idea of the tracer model is really tied to Fickian diffusion and components
with small concentration diffusing in a bulk phase. Hence also it's name.
The Maxwell-Stefan test actual models 3 gaseous components where none of
the components is really the main phase but most importantly there is no
bulk phase with a main component that is not considered in the equations.
Thus the 1pnc model is the right choice for such a test.
This changes the test to derive from the 1pnc model. As for the given
setting with only Neumann boundaries the 1pnc wouldn't convergence for
a compressible fluid, the Ideal gas law is used to model a
compressible gas.

The same test actually exists for the 1pnc model already in 1pnc/implicit/1p3c.
Thus this commit simply removes the tracer test.