[bin][extractlinedata] do not overwrite previous .csv files
When multiple vtk files were given as arguments to the script together
with a base name for the output .csv files, a single csv file was
created that contained the data of the last given vtk file. This commit
adds an index to the file names in case multiple vtk files are provided.