[md][embedded] Restructure 1d3d couplmanagers. Replace enums by tags.
The enums provide a convenient way to select the type of 1d3d coupling mode.
However they are not extensible. Adding a new mode requires changing Dumux core code.
With the new comparable and named tags, we have the same features but in
an extensible way. Every coupling manager implementation can define their own
(unique) tag and specialize the common class with this tag. The coupling mode
can be queried as before, and it is still easy to write code that works
with several coupling modes.

This commit also puts all coupling mode specializations in their own headers.