[md][embedded][average] Add surface factor correcting source for boundary-intersecting segments
If segments intersect the boundary of the 3d domain not all of the surface of the tubular inclusion
will couple with the 3d domain. The surface factor accounts for the reduction in surface area for
cylinder surfaces that are partly inside partly outside the 3d domain. This measure has proven
to slightly reduce the difference at boundaries between different coupling schemes.