[newton][varsbackend] add backends for vectors & generic variables
This commit also removes the interfaces related to the construction of
vectors for the solver, as this is currently done in Detail::assign in
newtonsolver.hh. The same holds for maxRelativeShift(), which is also
currently implemented in the Detail namespace in newtonsolver.hh-

In the future, we could think about outsourcing putting these two
functions into the backend.