[linear][test] Add a class for the Trilinos backend
Currently, the solve method iterates through the matrix and vector
entries of the linear system. This constitutes the basis for copying
the entries to corresponding Trilinos analogues. Include also the
necessary Trilinos headers for matrices and vectors.

Add a test to the incompressible 1p scenario. The test currently
still fails because the linear solvers doesn't converge. The
corresponding CMakeLists.txt still contains an architecture-dependent
path to the MPI C++ library that is not set by default.