[bin][createparamlist] print multiple differing occurrences
This also changes the result as now the first default value that is not
None is chosen in the table. Thus, entries that were empty before might
now contain a default value. Also, we now use "-" instead of empty
colummns for parameters where no default value is specified anywhere in
the code.