[TEMP][context] hard-code element solution
In dumux we obtain element solutions usually by calling the free
function elementSolution(...). This is called from within the assembly,
but there are no means to inject a different element solution type by
other means than providing an overload for elementSolution() for a
custom gridGeometry type or so. Therefore, this commit hardcodes element
solution in LocalContext such that the user does not have to specify the
type. Moreover, it introduces a convenience alias in order to be able to
deduce the element solution type within the context.

TODO:
- Alternatively we could make it a property
- rethink/discuss the convenience alias