[nonisothermal] compute contribution of p.div(v) depending on discretization
By assuming a constant cell pressure, the contribution of p.div(v) is
calculated by summing p.v.n over all scvfs of one scv. As p corresponds to
the scv, one such term (p.v.n)_scvf is usually different for the two scvs
sharing the scvf.
By default, each term is calculated separately for each scv and scvf by
means of the routine heatConvectionFlux. For box, symmetric fluxes are
assumed and the heat convection for one scvf is only calculated once for
both participating scvs. Therefore, the calculation of p.div(v) for box
is outsourced to a function computeVolumeWork which is called as part of
evalSource.