[navierstokes][problem] Do not expose discretization methods in problem
This created some unresolved dependenices due to the cyclic interdependency
of problem and coupling manager. This worked before because the discretizaiton
method was always one specific type. With tags the actual type has to be known,
so both problems and the coupling manager have to be instatiated before this
information can be accessed.