[couplingmanager] Use raw pointer to sub-problems
The pointer were previously managed by weak_ptr. As it turns out,
the interface obtaining the sub-problems is called very often in typical applications,
and can impact runtime. We therefore make it more efficient by using raw pointers
and refrain from using exception handling. In the case that the sub-problem pointers are not set,
an assert will be triggered in debug mode.