[test][1pnc] Improve Maxwell-Stefan test
Make test more robust by

* fixing output interval so always the same number of outputs are written
* use fixed time step size

Decrease runtime by only using 1 cell in y-direction.
We test in 2D to test the model in 2D but the specific setup is
one-dimensional so the added cells can be spared.

The new reference solution was created with a timestep
of 5s. The test uses a timestep of 25s which still
reproduces and accurate enough solution so the
comparison passes.