[test][md][facet] Increase test robustness by always comparing against last time step
The number of time steps may vary. With this we only write intial and last time step
and thus always compare to the solution corresponding to the final time. This makes
the test insensitive to small changes in the number of time steps (which can
occur due to rounding differences on different machines/compilers)