[porenetwork][common] Fix snap-off calculations for various cross-section by considering corner half angle
Snap-off happens in throats which have corners. To calculate the pc-snap-off, corner half angle of the throat is needed.
Before this change pc-snapoff was calculated only based on square cross section.
We need to include corner half angle in the calculation to be able to calculate pc snap-off for other cross sections than square.