[box] Deprecate scv/scvf.corner(i) interface. Use fvGeometry.geoemtry(scv/scvf).corner(i)
This may entail a runtime penalty in cases where this interface was used efficiently.
We accept this runtime penalty because it allows us to get rid of a big memory overhead
in all use cases (when caching is enabled). Geometry information of the scv/scvf will
seldom be required for standard assembly routines that usally rely on volumes and
normal information only.