[disc] add iterator range over elem-local-index/intersection pairs
This is helpful on network/surface grids, where multiple intersections
(one for each neighbor) are defined at the same geometric location. For
cell-centered finite-volume schemes, one needs to identify all neighbors
of a particular face, and therefore, this iterator range comes in handy
as it maps all geometrically coinciding intersections to the same index.