[extract module] add option to include GPLv3 license to new module
So far there was no option to automatically add a a GPLv3 license to a new module that was created with the 'extract_as_new_module.py' script.
This led to many modules being published without proper licensing. This extension adds the option to automatically add said license interactively during the module setup with the script.
Information about licensing is also added to the README.md file of the module.
The license text can be found in a file within the 'bin/license_for_module_extraction/' directory.