[assembler][pdesolver] Distinguish Resiudal and Solution types and set residual to dune native type
The solution vector type can contain switchable privars (vectors with additional information) as blocks.
The residual type now uses native dune types as this is the only thing required in the solver and assembler.
This change is not backwards-compatible but does not require changes in common main files.
If someone implemented their own assembler, solver, or nonlinear solver they might be affected by this change.
This commit then also shows how to make these components compatible with the change.