[disc] Deprecate scv/scvf.geometry/corner
The implementation in the fvgeometry is still using the scv/scvf interface for now
but all other users will get a deprecation warning. This should be enough to then
be able to replace the implementation in the fvGeometry and remove the interface
of the scv/scvf at the same time sometime after release 3.7.