[ff][bugfix] Use mutable cache instead of singleton
The singleton approach fail as soon as one has two instances
of a coupling manager in the same program. The occurring error
is very hard to debug. Since the singleton failed to make the
coupling manager thread safe, we go back to mutable caches.

Note that theoretically we could work with locks but this
has been tested to introduce large overhead. Global caches
will need to be eliminated eventually. In the meantime, we
at least make it work for multiple couplingmanager instances.