[timeloop] Add function to compute the time step size to the next check point
This is needed to be able to make adjustments to the time step size in the main
file in at least one corner case: we can get very close to the check point but can't
reach it because the set maximum time step size does not allow so. This results
in a tiny time step that might be unacceptable in a given application.