[test] Update outdated/invalid reference file
The Float32 field S_napl contained values like 1e-43 that cannot be
represented as float. This likely resulted from a writer that actually
wrote doubles. This issue seems to be fixed in newer version of the
Dune VTK writer and all written values fall within the float range.
Hence, we simply update the reference file with the file after read/write.
This solves an issue with the new gridformat reader which is more
strict when parsing floating point values than the Dumux::VTKReader based
on tinyxml2.